# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Skoli api",
        "version": "1.0.0",
        "x-logo": {
            "url": "https://via.placeholder.com/190x90.png?text=L5-Swagger"
        }
    },
    "paths": {
...
        "/api/profile/{id}": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "summary": "Get user data",
                "operationId": "dc381917fc0ff732960235c596a202d1",
                "responses": {
                    "200": {
                        "description": "Gets user data for the profile page",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": "1"
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Janko hrasko"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "janko@email.com"
                                                },
                                                "age": {
                                                    "type": "integer",
                                                    "example": 20
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "Profile"
                ],
                "summary": "Updates the profile",
                "operationId": "74d5575b9329d290b47768b88d73766a",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "age": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "petko hrasko",
                                    "age": 12
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "We store the rating",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "message": {
                                                    "type": "string",
                                                    "example": "Success"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}